requirements.txt
fastapi==0.118.0
python-telegram-bot==22.5
sqlalchemy==2.0.43
alembic==1.16.5
asyncpg==0.30.0
pydantic==2.11.10
uvicorn==0.37.0